Encapsulate user_links inside calculate_reputation function
This variable is not used outside of `calculate_reputation`, and there
is no reason to rely on the external call site to pre-calculate it, so
this patch localizes it inside `calculate_reputation` and replaces the
old `init_user_links` function with a new `calculate_user_links` which
generates and returns the vector rather than just filling its values.

This does not appear to have any performance impact.